merged

+++ hub/src/SshKeysController.cpp
@@ -180,7 +180,7 @@ void CSshKeysController::send_data_to_hub() {
       QString key_name = m_lst_key_files[col];
 
       for (int32_t row = 0; row < m_rows; ++row) {
-        if (m_current_ke_matrix[row].size() <= col) {
+        if (m_current_ke_matrix[row].size() <= (size_t)col) {
           continue;
         }
 
@@ -303,7 +303,7 @@ QStringList CSshKeysController::keys_in_environment(
   bool found = false;
 
   SynchroPrimitives::Locker lock(&csRbm);
-  for (row = 0; row < m_lst_healthy_environments.size(); ++row) {
+  for (row = 0; row < (int32_t)m_lst_healthy_environments.size(); ++row) {
     if (m_lst_healthy_environments[row].id() != env_id) continue;
     found = true;
     break;
